week6-security/requirements.txt
fastapi
uvicorn
pydantic
PyJWT